requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
autopep8==1.5.3
cachelib==0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
filelock==3.0.12
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==1.4.6
parsel==1.6.0
pem==20.1.0
Protego==0.1.16
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
Scrapy==2.1.0
scrapy-rotating-proxies==0.6.2
scrapy-splash==0.7.2
scrapy-user-agents==0.1.1
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.17
tabulate==0.8.7
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typing==3.7.4.1
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
virtualenv==20.0.23
w3lib==1.22.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
